Refactored interface to support asynchronous event logging
Right now, only the interface part understands the concept of concurrent timelines. The respective information is transmitted to the main beliefstate system, but is not yet properly interpreted (and will not result in the expected results if you use it for concurrent logging purposes!). Nested contexts and their annotations work as expected.